+++ extensions/mssql/src/azure/accountStore.ts
@@ -8,7 +8,8 @@ import * as vscode from "vscode";
 import * as Constants from "../constants/constants";
 import * as Loc from "../constants/locConstants";
 import { IAccount } from "../models/contracts/azure";
-import { ILogger, logger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
+import { logger } from "../models/logger";
 import { Deferred } from "../protocol";
 import { getErrorMessage } from "../utils/utils";
 import VscodeWrapper from "../controllers/vscodeWrapper";
