+++ extensions/mssql/src/azure/azureController.ts
@@ -16,7 +16,8 @@ import { getCloudProviderSettings } from "./providerSettings";
 import VscodeWrapper from "../controllers/vscodeWrapper";
 import { ConnectionProfile } from "../models/connectionProfile";
 import { AzureAuthType, IAADResource, IAccount, ITenant, IToken } from "../models/contracts/azure";
-import { ILogger, Logger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
+import { Logger } from "../models/logger";
 import { INameValueChoice, IPrompter, IQuestion, QuestionTypes } from "../prompts/question";
 import { AccountStore } from "./accountStore";
 import { ICredentialStore } from "../credentialstore/icredentialstore";
