+++ extensions/mssql/src/fabric/fabricHelper.ts
@@ -21,7 +21,8 @@ import { AxiosResponse } from "axios";
 import { getErrorMessage } from "../utils/utils";
 import { Fabric as Loc } from "../constants/locConstants";
 import { getCloudProviderSettings } from "../azure/providerSettings";
-import { ILogger, logger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
+import { logger } from "../models/logger";
 
 export class FabricHelper {
     static getFabricApiUriBase(): vscode.Uri {
