+++ extensions/mssql/src/http/httpClient.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from "vscode";
 import * as LocalizedConstants from "../constants/locConstants";
-import { ILogger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
 import { getErrorMessage } from "../utils/utils";
 import {
     HttpClientCore,
