+++ extensions/mssql/src/languageservice/interfaces.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as tmp from "tmp";
-import { ILogger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
 import * as vscodeMssql from "vscode-mssql";
 
 export interface IStatusView {
