+++ extensions/mssql/src/mssqlProtocolHandler.ts
@@ -10,7 +10,8 @@ import SqlToolsServiceClient from "./languageservice/serviceclient";
 import { CapabilitiesResult, GetCapabilitiesRequest } from "./models/contracts/connection";
 import { IConnectionProfile } from "./models/interfaces";
 import { AuthenticationType } from "./sharedInterfaces/connectionDialog";
-import { ILogger, logger } from "./models/logger";
+import { ILogger } from "./sharedInterfaces/logger";
+import { logger } from "./models/logger";
 import VscodeWrapper from "./controllers/vscodeWrapper";
 import MainController from "./controllers/mainController";
 import { cmdAddObjectExplorer } from "./constants/constants";
