+++ extensions/mssql/src/notebooks/notebookQueryExecutor.ts
@@ -27,7 +27,7 @@ import { QueryDisposeRequest, QueryDisposeParams } from "../models/contracts/que
 import { QueryCancelRequest, QueryCancelParams } from "../models/contracts/queryCancel";
 import { IDbColumn, IResultMessage } from "../models/interfaces";
 import { Deferred } from "../protocol";
-import { ILogger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
 
 export interface NotebookResultSetData {
     columnInfo: IDbColumn[];
