+++ extensions/mssql/src/notebooks/sqlNotebookController.ts
@@ -12,7 +12,8 @@ import * as LocalizedConstants from "../constants/locConstants";
 import ConnectionManager from "../controllers/connectionManager";
 import { ConnectionSharingService } from "../connectionSharing/connectionSharingService";
 import * as Utils from "../models/utils";
-import { ILogger, Logger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
+import { Logger } from "../models/logger";
 import { NotebookConnectionManager } from "./notebookConnectionManager";
 import { NotebookCodeLensProvider } from "./notebookCodeLensProvider";
 import { NotebookBatchResult } from "./notebookQueryExecutor";
