+++ extensions/mssql/src/objectExplorer/objectExplorerDragAndDropController.ts
@@ -9,7 +9,8 @@ import { TreeNodeInfo } from "./nodes/treeNodeInfo";
 import { ObjectExplorerUtils } from "./objectExplorerUtils";
 import { ConnectionNode } from "./nodes/connectionNode";
 import { ConnectionGroupNode } from "./nodes/connectionGroupNode";
-import { ILogger, logger } from "../models/logger";
+import { ILogger } from "../sharedInterfaces/logger";
+import { logger } from "../models/logger";
 import VscodeWrapper from "../controllers/vscodeWrapper";
 import { getErrorMessage } from "../utils/utils";
 import { ConnectionStore } from "../models/connectionStore";
